# apps/backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JobBoard",
    "description": "JobBoard API documention",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000/api/v1",
      "description": "Local "
    },
    {
      "url": "http://dev.joboard.io/api/v1",
      "description": "Dev "
    },
    {
      "url": "http://stage.joboard.io/api/v1",
      "description": "Stage "
    }
  ],
  "tags": [
    {
      "name": "locations",
      "description": "Location api endpoint"
    },
    {
      "name": "categories",
      "description": "Categories api endpoint"
    },
    {
      "name": "authentication",
      "description": "Authentication api endpoint"
    }
  ],
  "paths": {
    "/authentication/signin": {
      "post" : {
        "tags": [
          "authentication"
        ],
        "summary": "Signin",
        "description": "Signin",
        "requestBody": {
          "description": "Signin",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "johndoe@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Login successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "access_token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY2NTE0NjBjNzU3YjFjNDI2NzZmZjM2ZiIsImlhdCI6MTcxNzAxMjk1NCwiZXhwIj"
                        },
                        "redirect_identifier": {
                          "type": "string",
                          "example": "user"
                        }
                      }
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": ["email should not be empty", "email must be an email", "password should not be empty"]
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }
    },
    "/authentication/signup": {
      "post" : {
        "tags": [
          "authentication"
        ],
        "summary": "Signup",
        "description": "Signup",
        "requestBody": {
          "description": "Signup",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Mitchell Skiles"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@gmail.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "0111045783"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Login successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "access_token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY2NTE0NjBjNzU3YjFjNDI2NzZmZjM2ZiIsImlhdCI6MTcxNzAxMjk1NCwiZXhwIj"
                        },
                        "redirect_identifier": {
                          "type": "string",
                          "example": "user"
                        }
                      }
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": ["name should not be empty", "email should not be empty", "email must be an email", "The password is too weak. Please choose a stronger one.", "Password must be at least 6 characters long", "password should not be empty"]
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }
    },
    "/locations": {
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "Get all locations",
        "description": "Get all locations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": null
                    },
                    "data": {
                      "type": "array",
                      "items" : {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "66066a1123cb331bef59d690"
                          },
                          "name": {
                            "type": "string",
                            "example": "California"
                          },
                          "status": {
                            "type": "string",
                            "example": "InActive"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-03-29T07:13:16.528Z"
                          }
                        } 
                      }
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "post" : {
        "tags": [
          "locations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a  location",
        "description": "Create a location",
        "requestBody": {
          "description": "Create a new location",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bexley"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Location has been created successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "name must be longer than or equal to 1 characters"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }
    },
    "/locations/{id}": {
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "Get a location",
        "description": "Get a location by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "location id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": null
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66066a1123cb331bef59d690"
                        },
                        "name": {
                          "type": "string",
                          "example": "California"
                        },
                        "status": {
                          "type": "string",
                          "example": "InActive"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-03-29T07:13:16.528Z"
                        }
                      } 
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "put" : {
        "tags": [
          "locations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a  location",
        "description": "Update a location",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "location id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bexley"
                  },
                  "status": {
                    "type": "string",
                    "example": "Active"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Location has been updated successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "name must be longer than or equal to 1 characters"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "delete": {
        "tags": [
          "locations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete location",
        "description": "Delete a location by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "location id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Location has been deleted successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }

    },
    "/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get all categories",
        "description": "Get all categories",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": null
                    },
                    "data": {
                      "type": "array",
                      "items" : {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "66066a1123cb331bef59d690"
                          },
                          "name": {
                            "type": "string",
                            "example": "Category AA"
                          },
                          "status": {
                            "type": "string",
                            "example": "InActive"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-03-29T07:13:16.528Z"
                          }
                        } 
                      }
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "post" : {
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a  category",
        "description": "Create a category",
        "requestBody": {
          "description": "Create a new category",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bexley"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Location has been created successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "name must be longer than or equal to 1 characters"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get a category",
        "description": "Get a category by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "category id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": null
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "66066a1123cb331bef59d690"
                        },
                        "name": {
                          "type": "string",
                          "example": "Category AA"
                        },
                        "status": {
                          "type": "string",
                          "example": "InActive"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-03-29T07:13:16.528Z"
                        }
                      } 
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "put" : {
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a  category",
        "description": "Update a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "category id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Bexley"
                  },
                  "status": {
                    "type": "string",
                    "example": "Active"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Category has been updated successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "422": {
            "description": "Validtion Errors",
             "content": {
               "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 422
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "name must be longer than or equal to 1 characters"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  } 
                }
              }
             }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete category",
        "description": "Delete a category by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "category id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Category been deleted successfully"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  } 
                }
              }
            }
          },
          "400": {
            "description": "Error operation",
             "content": {
               "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
             }
          }
        }
      }

    }
  },
  "components": {
    "schemas": {
      "ErrorApiResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Something went wrong, Please try again later"
          },
          "data": {
            "type": "string",
            "example": null
          }
        }
      }
    },

    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}